Fix boundary parts + safety radii for pair construction

+++ src/hydro/Shadowswift/hydro_part.h
@@ -253,7 +253,7 @@ struct part {
 
     /*! Flags indicating to which neighbouring cells this particle has already
      * been added. */
-    int delaunay_flags;
+    long delaunay_flags;
 
   } geometry;
